+++ tests/test_logging.py
@@ -1,7 +1,7 @@
 #
 # MIT License
 #
-# (C) Copyright 2019-2020 Hewlett Packard Enterprise Development LP
+# (C) Copyright 2019-2020,2023 Hewlett Packard Enterprise Development LP
 #
 # Permission is hereby granted, free of charge, to any person obtaining a
 # copy of this software and associated documentation files (the "Software"),
@@ -40,10 +40,13 @@ class TestLogging(unittest.TestCase):
         Sets up by patching getLogger to return a logger of our own and by
         patching get_config_value to return values from a dict.
         """
+        # In the logging module, getLogger is called twice, once to get the main
+        # logger for SAT and once to get the logger for the csm_api_client library.
         self.logger = logging.getLogger(__name__)
+        self.sub_logger = logging.getLogger('submodule')
         self.mock_get_logger = mock.patch(
             'sat.logging.logging.getLogger',
-            return_value=self.logger
+            side_effect=(self.logger, self.sub_logger)
         ).start()
 
         config_values = self.config_values = {
@@ -123,7 +126,9 @@ class TestLogging(unittest.TestCase):
         configure_logging()
 
         # This should have gotten the logger named 'sat'
-        self.mock_get_logger.assert_called_once_with('sat')
+        self.mock_get_logger.assert_any_call('sat')
+        # And also the logger named 'csm_api_client'
+        self.mock_get_logger.assert_any_call('csm_api_client')
 
         log_dir = os.path.dirname(self.config_values['logging.file_name'])
         mock_makedirs.assert_called_once_with(log_dir, exist_ok=True)
@@ -152,7 +157,8 @@ class TestLogging(unittest.TestCase):
         configure_logging()
 
         # This should have gotten the logger named 'sat'
-        self.mock_get_logger.assert_called_once_with('sat')
+        self.mock_get_logger.assert_any_call('sat')
+        self.mock_get_logger.assert_any_call('csm_api_client')
 
         # Exactly one handler of type StreamHandler should have been added.
         self.assertEqual(len(self.logger.handlers), 1)